fix: rewrite to avoid code duplication + fix mistake in documentation
